requirements/local.txt
# Local development dependencies go here
-r base.txt
coverage==4.2
Sphinx
django-extensions==1.7.4
Werkzeug==0.11.11

# django-debug-toolbar that works with Django 1.5+
django-debug-toolbar==1.5
